Add soft limit option extension
Function 'toggle_softlims.m' is updated to allow not just soft limits on
line rating but on votlage magnitude, generator limits, and angle
differences as well. Additionally, in stead of always creating a soft
limit bounded by infinity, the options to specify a fixed value or a
fraction of the hard limit is implemented.

Tests in 't/t_opf_softlims.m' have been updated in accordance witht the
new data structure. All previous tests have been kept and several new
added for the different limit types.

Utility function 'calc_branch_angle.m' is added, which simply calculates
the angle differences accross active branches.

NOTE 1: In keeping with the previous implementation the print callback
function selects all elements where the soft limit is used (non-zero
overload) OR where the shadow price is non-zero. As a result the
softlimits section prints some items, where the softlimits are
technically not active.

NOTE 2: The softlims_defaults function is responsible for setting up
much of the softlims structure. This includes creating an appropriately
sized upper bound 'ub' vector for the slack variable. If an upper bound
vector is specified, it is used as is. It is therefore NOT recommended
to set up softlims on an mpc, run opf, save the results, modify the
softlims structure in the results, and rerun opf. Instead, the original
mpc should be altered.

 Changes to be committed:
	new file:   lib/calc_branch_angle.m
	modified:   lib/t/t_opf_softlims.m
	modified:   lib/toggle_softlims.m