Integration of opf.softlims.default option
The syntax for activating softlims now accepts an optional argument
which should be the matpower options structure, i.e:
	mpc = toggle_softlims(mpc, 'on', mpopt)
This adds `mpopt` to as an optional argument to the `ext2int` callback.
If `mpopt.opf.softlims.default = 0`, any unspecified softlimits are
neglected when setting up the basic softlims structure. That is they are
set up with `type='none'`.

Test function `t_opf_softlims.m` is updated with a few tests to make
sure this new option works as expected.

 Changes to be committed:
	modified:   t/t_opf_softlims.m
	modified:   toggle_softlims.m