Fix error in gen ordering in ext2int ordering
The ext2int function reorders the gen matrix, which needs to then be
reflected in the softlims.prop.idx vector. This was done with,
	s.prop.idx = s.prop.idx(o.gen.e2i)
where o is the order structure.
However, in cases where not all generators have softlimits applied the
indices of o.gen.e2i could be out of bounds.
This was fixed to be,
	s.prop.idx = s.prop.idx(o.gen.e2i(s.prop.idx))
where indexing permutation o.gen.e2i with s.prop.idx should alleviate
the issue.