Changes to softlims names and default behavior
field names were changed from 'type' to 'hl_mod', in addition field
'hl_val' is added that indicates what the NEW hard limit will be.
The scaling modification was changed to be hl_val*original limit instead
of the previous implementation where it directly indicated what the
upperbound on the slack variable will be.

rval for PMIN and VMIN is now -Inf instead of 0 as previously

Defaults were changed to be 'remove' (previously called 'unbnd') for all
limits except VMIN and PMIN. VMIN uses 'replace' and sets the new hard
limit to 0. While PMIN uses 'replace' and sets the new hard min to 0 for
normal generators (PMIN > 0) and -Inf for generators with PMIN < 0 and
PMAX > 0 (dispachable loads with PMIN <0 and PMAX = 0 are ignored by
softlims).

hl_mod 'shift' was also added, which shifts the hard limit by a fixed
amount (direction is dependent on the limit).

The tests were updated to reflect these changes as well as a couple
tests added to make sure that the upper limits are calculated correctly
given the different hl_mod options.