Fix index handling for bus matrix
The idx vector for VMIN and VMAX is supposed to contain bus numbers.
Internally this was converted to row indices but this conversion was not
properly undone in the int2ext callback.

This has now been fixed a long with the necessary changes to the print
callback and some tests that verify this is doing what it is supposed
to.