Check for opf success in int2ext callback
The success field is added to the results structure now just before the
int2ext function is called in `opf`.
This way, the various callbacks can utilize this.
Specifically, the toggle_softlims int2ext callback does not perform a
check that was causing uncessary fatal erros if the success field is 0.